Pull out test helpers to its own module
